Refactored type definitions to @types and removed babel
In TypeScript 2.0 typings is moved to node_modules/@types - refactored application to support this. As part of this change
React typings were changed to have generics added to the SyntheticEvent interface. Code was refactored to satisfy the
typings requirement.

Also removed all babel references since we are not using ES7 features.